Introduce generic HttpTileRetriever and CachedHttpTileRetriever to simplify feature tile server implementations. Refactor CachedOsmTileRetriever to use CachedHttpTileRetriever.
